+++ R/r2f-aab-core.R
@@ -249,8 +249,12 @@ lang2fortran <- r2f <- function(
         # logicals passed via the bind(c) interface are stored as integer(0/1)
         # and must be "booleanized" for Fortran logical operations.
         s <- paste0("(", s, "/=0)")
+        out <- Fortran(s, value = if (inherits(val, Variable)) val else NULL)
+        attr(out, "logical_booleanized") <- TRUE
+        out
+      } else {
+        Fortran(s, value = if (inherits(val, Variable)) val else NULL)
       }
-      Fortran(s, value = if (inherits(val, Variable)) val else NULL)
     },
 
     ## handling 'object' and 'closure' here are both bad ideas,